entities 2.0
entities are now:
- affected by size
- have an "DNA"
- lose power by moving
- gain power by eat other entities
- do not kill their own clones
- stop moving if power is 0
- random color
- different speed

+++ Assets/Scripts/DisplayManager.cs
@@ -32,13 +32,13 @@ public class DisplayManager : MonoBehaviour
         GameObject lowPowerEntity = null;
         if (timeLastUpdate >= 0.1f)
         {
-            int highPower = 0;
-            int lowPower = 1000;
+            float highPower = 0;
+            float lowPower = 1000;
             evos = GameObject.FindGameObjectsWithTag("Entity");
             foreach (GameObject checkEntity in evos)
             {
 
-                int checkPower = checkEntity.GetComponent<Entity>().power;
+                float checkPower = checkEntity.GetComponent<Entity>().power;
 
                 if (checkPower > highPower)
                 {
